requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
matplotlib>=3.6.0
requests>=2.28.0
python-fire>=0.5.0
python-dotenv>=1.0.0
openai